# backend  - 파일업로드기능 완료

+++ src/main/java/com/genesislabs/video/entity/FileUploadEntity.java
@@ -21,5 +21,5 @@ public class FileUploadEntity {
     @NotNull
     private String vf_filename;
     @NotNull
-    private Integer vf_filesize;
+    private Long vf_filesize;
 }
